requirements.txt
numpy
SoundCard
pywebview
winsdk